backend/requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
flask-marshmallow==1.3.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
marshmallow==4.0.0
marshmallow-sqlalchemy==1.4.2
PyMySQL==1.1.1
SQLAlchemy==2.0.40
typing_extensions==4.13.0
Werkzeug==3.1.3
cryptography
debugpy
flask-cors
flask-jwt-extended
